refactor(utils/auth-provider): revise auth provider logics
