+++ packages/mesh/app/login/_components/login-email-confirmation.tsx
@@ -1,6 +1,6 @@
 import { useLogin } from '@/context/login'
-import { sendEmailLink } from '@/hooks/use-auth-provider'
 import useCountdown from '@/hooks/use-countdown'
+import { sendEmailLink } from '@/utils/auth-provider'
 
 export default function LoginEmailConfirmation() {
   const { formData, setStep } = useLogin()
